partial changes for refactoring storage/analytics.
