+++ arangopipe/MANIFEST.in
@@ -1,1 +1,1 @@
-include arangopipe/arangopipe_config.yaml
+include arangopipe/arangopipe_storage/arangopipe_config.yaml
