+++ arangopipe/setup.py
@@ -11,7 +11,7 @@ README = (HERE / "README.md").read_text()
 # This call to setup() does all the work
 setup(
     name="arangopipe",
-    version="0.0.47.2",
+    version="0.0.47.4",
     description="package for machine learning meta-data management and analysis",
     long_description=README,
     long_description_content_type="text/markdown",
